pyproject.toml
[build-system]
requires = ["setuptools>=77", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shmpipeline"
version = "0.1.0"
description = "Process-based shared-memory compute pipelines built on pyshmem."
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.9"
license = "GPL-3.0-only"
license-files = ["LICENSE"]
authors = [
    { name = "Jacob Taylor" },
]
dependencies = [
    "numba>=0.60",
    "numpy>=1.26,<3",
    "pyshmem>=1.0.0",
    "PyYAML>=6.0",
]
keywords = ["shared-memory", "pipeline", "multiprocessing", "numba", "cuda"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3",
    "pytest-cov>=5",
    "ruff>=0.11",
]
gpu = [
    "torch>=2.2",
]
gui = [
    "PySide6>=6.8",
    "pyqtgraph>=0.14",
]

[project.scripts]
shmpipeline = "shmpipeline.cli:main"
shmpipeline-gui = "shmpipeline.gui.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["shmpipeline", "shmpipeline.*"]

[tool.ruff]
line-length = 79
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "I"]
ignore = ["E501"]